refactor de clases y packages para mantener uniformidad
